Allow removal of default status on a term
When an alert expires, it sets its level back to a default alert
level. If no alert level is set as the default, it clears the alert
entirely.

Previously, it was not possible to clear the default alert level,
so an expiring alert would always revert to the default level.

After this change, expiration works as expected.

Fixes #22